+++ frontend/fingertips-frontend/lib/chartHelpers/chartHelpers.test.ts
@@ -1,6 +1,7 @@
 import {
   generateSeriesData,
   sortHealthDataByDate,
+  sortHealthDataByYearDescending,
 } from '@/lib/chartHelpers/chartHelpers';
 
 const mockData = [
@@ -78,5 +79,65 @@ describe('generateSeriesData', () => {
       },
     ];
     expect(result).toEqual(mockSeriesData);
+  });
+});
+
+describe('sortHealthDataByYearDescending', () => {
+  it('should sort the healthcare data  in descending years', () => {
+    const mockData = [
+      {
+        areaCode: 'A1425',
+        areaName: 'area A1425',
+        healthData: [
+          {
+            count: 267,
+            lowerCi: 441.69151,
+            upperCi: 578.32766,
+            value: 703.420759,
+            year: 2004,
+            sex: 'Persons',
+            ageBand: 'All',
+          },
+          {
+            count: 389,
+            lowerCi: 441.69151,
+            upperCi: 578.32766,
+            value: 278.29134,
+            year: 2006,
+            sex: 'Persons',
+            ageBand: 'All',
+          },
+        ],
+      },
+    ];
+    const mockSortedData = [
+      {
+        areaCode: 'A1425',
+        areaName: 'area A1425',
+        healthData: [
+          {
+            count: 389,
+            lowerCi: 441.69151,
+            upperCi: 578.32766,
+            value: 278.29134,
+            year: 2006,
+            sex: 'Persons',
+            ageBand: 'All',
+          },
+          {
+            count: 267,
+            lowerCi: 441.69151,
+            upperCi: 578.32766,
+            value: 703.420759,
+            year: 2004,
+            sex: 'Persons',
+            ageBand: 'All',
+          },
+        ],
+      },
+    ];
+
+    const result = sortHealthDataByYearDescending(mockData);
+    expect(result).toEqual(mockSortedData);
   });
 });
