--- frontend/fingertips-frontend/lib/chartHelpers/scatterChartHelper.ts
@@ -1,10 +1,0 @@
-import { HealthDataForArea } from '@/generated-sources/ft-api-client';
-
-export function sortHealthDataByYearDescending(
-  data: HealthDataForArea[] = []
-): HealthDataForArea[] {
-  return data.map((item) => ({
-    ...item,
-    healthData: item.healthData.toSorted((a, b) => b.year - a.year),
-  }));
-}
